+++ python/cuml/neighbors/kneighbors_regressor_mg.pyx
@@ -16,8 +16,11 @@
 
 # distutils: language = c++
 
+import typing
+
 import numpy as np
 
+import cuml.internals
 from cuml.common.array import CumlArray
 from cuml.raft.common.handle cimport handle_t
 from cuml.common import input_to_cuml_array
@@ -68,9 +71,22 @@ class KNeighborsRegressorMG(KNeighborsMG):
         super(KNeighborsRegressorMG, self).__init__(**kwargs)
         self.batch_size = batch_size
 
-    def predict(self, data, data_parts_to_ranks, data_nrows,
-                query, query_parts_to_ranks, query_nrows,
-                ncols, n_outputs, rank, convert_dtype):
+    @cuml.internals.api_base_return_generic_skipall
+    def predict(
+        self,
+        data,
+        data_parts_to_ranks,
+        data_nrows,
+        query,
+        query_parts_to_ranks,
+        query_nrows,
+        ncols,
+        n_outputs,
+        rank,
+        convert_dtype
+    ) -> typing.Tuple[typing.List[CumlArray],
+                      typing.List[CumlArray],
+                      typing.List[CumlArray]]:
         """
         Predict outputs for a query from previously stored index
         and index labels.
@@ -93,9 +109,7 @@ class KNeighborsRegressorMG(KNeighborsMG):
         -------
         predictions : outputs, indices, distances
         """
-        out_type = self.get_out_type(data, query)
-
-        out_type = self.get_out_type(data, query)
+        self.get_out_type(data, query)
 
         input = self.gen_local_input(data, data_parts_to_ranks, data_nrows,
                                      query, query_parts_to_ranks, query_nrows,
@@ -148,10 +162,6 @@ class KNeighborsRegressorMG(KNeighborsMG):
             free(<void*>out_result_local_parts.at(i))
         free(<void*><uintptr_t>out_result_local_parts)
 
-        output = list(map(lambda o: o.to_output(out_type), output_cais))
-        output_i = list(map(lambda o: o.to_output(out_type),
-                            result['cais']['indices']))
-        output_d = list(map(lambda o: o.to_output(out_type),
-                            result['cais']['distances']))
-
-        return output, output_i, output_d
+        return output_cais, \
+            result['cais']['indices'], \
+            result['cais']['distances']